Font, Scrollbar, Course Search
The HTML for the search is all set, the CSS is almost there, and the JS needs a lot of work. The scroll is now a rounded rectangle. The font has been changed to Stonehill's standard font but I'm not totally sure if its taking effect on everything. It's def taking effect on the navbar.